packages/kharkiv-metro-bot/pyproject.toml
[project]
name = "kharkiv-metro-bot"
version = "0.1.0"
description = "Telegram Bot for Kharkiv Metro Route Planner"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.0.0",
    "python-dotenv>=1.0.0",
    "kharkiv-metro-core",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
metro-bot = "kharkiv_metro_bot.main:main_sync"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kharkiv_metro_bot"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
